Revert "Merge branch 'dev/blue/Readme' of https://github.com/Turma-IF-CIC-2025-2027/All-in into dev/blue/Readme"
This reverts commit 3e29dc1fcd7bef227259dbeea08ecd8c52869c3b, reversing
changes made to 55f3e043a5b67f0bb2fe80901390a4ac7e46ec4e.
