--- .env.example
@@ -1,10 +1,0 @@
-NODE_ENV=dev
-JWT_SECRET=abc123
-JWT_EXPIRES_IN=2h
-PORT=80
-CORS_ORIGIN=http://localhost:80
-DB_HOST=localhost
-DB_PORT=3306
-DB_USER=root
-DB_PASSWORD=
-DB_NAME=all_in_db
