+++ server.ts
@@ -1,7 +1,23 @@
-import app from './app';
+import "reflect-metadata";
+import express from 'express';
+import path from 'path';
+import cors from 'cors';
 import { AppDataSource } from './src/core/database/data-source'; //TypeORM config
+//import allRoutes from './modules/router'; //Routes
 
-const PORT = process.env.PORT || 80;
+const app = express();
+const PORT = process.env.PORT || 8080;
+
+//Global Middlewares
+app.use(cors());  
+app.use(express.json()); 
+app.use(express.urlencoded({ extended: true }));
+
+//Add Static Files
+app.use(express.static(path.join(__dirname, '../public')));
+
+//Routes Injection
+//app.use('/', allRoutes);
 
 //Initialize BD
 /*AppDataSource.initialize()
@@ -14,8 +30,11 @@ const PORT = process.env.PORT || 80;
     })
     .catch((error:any) => {
         console.error("Error connecting to the database", error);
-        process.exit(1);
     });*/
+
+app.get('/', (req, res) => {
+    res.send("Hello World!");
+});
 
 //Remove after this comment, just for testing while there is no DB
 app.listen(PORT, () => {
