--- src/modules/auth/views/login.html
@@ -1,1 +1,0 @@
-<h1>Login</h1>
