--- src/modules/auth/views/register.html
@@ -1,1 +1,0 @@
-<h1>Register</h1>
