+++ src/modules/example/example-routes.ts
@@ -1,0 +1,13 @@
+import { Router } from 'express';
+//import { ExampleController } from './ExampleController';
+
+const router = Router();
+//const ExampleController = new ExampleController(); // Instanciamos o controller
+
+//router.post('/register', (req, res) => authController.register(req, res));
+//router.post('/login', (req, res) => authController.login(req, res));
+//router.post('/logout', (req, res) => authController.logout(req, res));
+
+//router.get('/me', ExampleMiddleware, (req, res) => authController.getProfile(req, res));
+
+export { router as ExampleRoutes };
