--- src/modules/profile/profile-service.ts
@@ -1,7 +1,0 @@
-export class ProfileService {
-    async getProfile(userId: string): Promise<any> {
-        // Implement get profile logic here
-        // e.g., retrieve user information from database, etc.
-        return {};
-    }
-}
