+++ src/modules/router.ts
@@ -1,18 +1,8 @@
 import { Router } from 'express';
-import { AuthRoutes } from './auth/auth-routes';
-import { ProfileRoutes } from './profile/profile-routes';
-import { authMiddleware } from '../core/middlewares/auth-middleware';
+import { ExampleRoutes } from './example/example-routes';
 
 const router = Router();
 
-//Public Routes
-router.use('/auth', AuthRoutes);
-router.get('/', (req, res) => {
-    res.send("<h1>Hello World!</h1><a href='/auth/login'>Login here</a> | <a href='/auth/register'>Register here</a>");
-});
-
-//Protected Routes
-router.use(authMiddleware);
-router.use('/profile', ProfileRoutes);
+router.use('/example', ExampleRoutes);
 
 export default router;
