+++ tsconfig.json
@@ -8,12 +8,13 @@
     // Environment Settings
     // See also https://aka.ms/tsconfig/module
     "module": "commonjs",
-    "moduleResolution": "node10",
+    "moduleResolution": "node",
     "target": "ES2020",
     "esModuleInterop": true,
     // For nodejs:
-    "lib": ["esnext"],
-    "types": ["node"],
+    // "lib": ["esnext"],
+    // "types": ["node"],
+    // and npm install -D @types/node
 
     // Other Outputs
     "sourceMap": true,
@@ -40,16 +41,6 @@
     "noUncheckedSideEffectImports": true,
     //"moduleDetection": "force",
     "skipLibCheck": true,
-
-    // TypeORM 
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "strictPropertyInitialization": false,
-    
-
-    // Paths
-    "baseUrl": "./",
-    "typeRoots": ["./node_modules/@types", "./src/types"]
   },
     "include": ["src/**/*" , "server.ts"]
 }
